requirements.txt
crewai
crewai[tools]
html2text
markdown
googlesearch-python
requests
beautifulsoup4